Move `concretization_version_order` to `package_base`
