+++ lib/spack/spack/solver/versions.py
@@ -2,24 +2,6 @@
 #
 # SPDX-License-Identifier: (Apache-2.0 OR MIT)
 import enum
-from typing import Tuple, Union
-
-from spack.version import GitVersion, StandardVersion
-
-
-def concretization_version_order(version_info: Tuple[Union[GitVersion, StandardVersion], dict]):
-    """Version order key for concretization, where preferred > not preferred,
-    not deprecated > deprecated, finite > any infinite component; only if all are
-    the same, do we use default version ordering."""
-    version, info = version_info
-    return (
-        info.get("preferred", False),
-        not info.get("deprecated", False),
-        not isinstance(version, GitVersion),
-        not version.isdevelop(),
-        not version.is_prerelease(),
-        version,
-    )
 
 
 class Provenance(enum.IntEnum):
